+++ Changes.md
@@ -14,6 +14,7 @@ Fixes
 API
 ---
 
+- CodeWidget : Added new class for displaying/editing code.
 - PlugLayout : Added a warning for plugs that reference activators that have not been registered.
 - ImageTestCase : Added `ignoreChannelNamesOrder` keyword argument to `assertImagesEqual()`. This defaults to `False`, maintaining the previous behaviour.
 - MultiLineTextWidget : Added `cursorBound()` method.